refactor: 프로필 관련 핸들러를 MemberController로 이동
