+++ src/main/java/com/b4f2/pting/controller/MemberController.java
@@ -3,7 +3,9 @@ package com.b4f2.pting.controller;
 import jakarta.validation.Valid;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -20,10 +22,15 @@ import com.b4f2.pting.domain.School;
 import com.b4f2.pting.dto.CertificationRequest;
 import com.b4f2.pting.dto.CertificationResponse;
 import com.b4f2.pting.dto.CertificationVerifyRequest;
+import com.b4f2.pting.dto.ChangeProfileDescriptionRequest;
+import com.b4f2.pting.dto.ChangeProfileImageUrlRequest;
+import com.b4f2.pting.dto.ChangeProfileNameRequest;
 import com.b4f2.pting.dto.GamesResponse;
+import com.b4f2.pting.dto.ProfileResponse;
 import com.b4f2.pting.dto.SchoolResponse;
 import com.b4f2.pting.facade.CertificationService;
 import com.b4f2.pting.service.GameService;
+import com.b4f2.pting.service.ProfileService;
 import com.b4f2.pting.service.SchoolService;
 
 @RestController
@@ -34,6 +41,7 @@ public class MemberController {
     private final CertificationService certificationService;
     private final SchoolService schoolService;
     private final GameService gameService;
+    private final ProfileService profileService;
 
     @PostMapping("/me/school/{schoolId}")
     public ResponseEntity<SchoolResponse> selectSchool(@Login Member member, @PathVariable Long schoolId) {
@@ -69,5 +77,41 @@ public class MemberController {
             return ResponseEntity.ok(gameService.findGamesByMember(member));
         }
         return ResponseEntity.ok(gameService.findGamesByMemberAndGameStatus(member, query));
+    }
+
+    @GetMapping("/me/profile")
+    public ResponseEntity<ProfileResponse> getMyProfile(@Login Member member) {
+        ProfileResponse profileResponse = profileService.getProfile(member.getId());
+        return ResponseEntity.ok(profileResponse);
+    }
+
+    @GetMapping("/{memberId}/profile")
+    public ResponseEntity<ProfileResponse> getProfile(@PathVariable Long memberId) {
+        ProfileResponse profileResponse = profileService.getProfile(memberId);
+        return ResponseEntity.ok(profileResponse);
+    }
+
+    @PatchMapping("/me/profile/name")
+    public ResponseEntity<ProfileResponse> updateName(
+            @Login Member member, @Validated @RequestBody ChangeProfileNameRequest request) {
+        ProfileResponse profileResponse = profileService.updateName(member.getId(), request.name());
+
+        return ResponseEntity.ok(profileResponse);
+    }
+
+    @PatchMapping("/me/profile/description")
+    public ResponseEntity<ProfileResponse> updateDescription(
+            @Login Member member, @Validated @RequestBody ChangeProfileDescriptionRequest request) {
+        ProfileResponse profileResponse = profileService.updateDescription(member.getId(), request.description());
+
+        return ResponseEntity.ok(profileResponse);
+    }
+
+    @PatchMapping("/me/profile/image-url")
+    public ResponseEntity<ProfileResponse> updateImageUrl(
+            @Login Member member, @Validated @RequestBody ChangeProfileImageUrlRequest request) {
+        ProfileResponse profileResponse = profileService.updateImageUrl(member.getId(), request.imageUrl());
+
+        return ResponseEntity.ok(profileResponse);
     }
 }
